+++ client/src/main/java/org/evosuite/testsmells/smells/EagerTest.java
@@ -43,7 +43,7 @@ public class EagerTest extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
 
         Statement currentStatement;
