+++ client/src/main/java/org/evosuite/testsmells/smells/MysteryGuest.java
@@ -20,9 +20,9 @@ public class MysteryGuest extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
+        long count = 0;
 
         Statement currentStatement;
