+++ client/src/main/java/org/evosuite/testsmells/smells/ObscureInlineSetup.java
@@ -34,9 +34,9 @@ public class ObscureInlineSetup extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
+        long count = 0;
 
         Statement currentStatement;
