+++ client/src/main/java/org/evosuite/testsmells/smells/Overreferencing.java
@@ -49,7 +49,7 @@ public class Overreferencing extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
         Statement currentStatement;
