+++ client/src/main/java/org/evosuite/testsmells/smells/TestCodeDuplication.java
@@ -10,9 +10,9 @@ public class TestCodeDuplication extends AbstractTestCaseSmell {
     }
 
     @Override
-    public double computeNumberOfTestSmells(TestChromosome chromosome) {
+    public long computeNumberOfTestSmells(TestChromosome chromosome) {
         int size = chromosome.size();
-        int count = 0;
+        long count = 0;
 
         int dist;
         double similar;
